Made Router and Manager into classes
