+++ manager.cc
@@ -1,62 +1,18 @@
-#include <stdio.h>
-#include <string.h>
-#include <iostream>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include<vector>
-#include <sys/wait.h>
-#include <unistd.h>
-#include <fstream>
-#include <fcntl.h>
-#include <sstream>
-
 #include "project3.h"
 
-using namespace std;
-
-int children;
 
 void parentFunction(string in) {
-    int fd, children;
-    string line;
-    vector<string> lines;
-    fd = open("test.txt", O_CREAT|O_TRUNC|O_WRONLY, 0666);
-    ifstream input (in);
 
-    if(input.is_open()) {
+    Manager manager(in);
+    manager.SpawnRouters();
 
-        // children holds how many forks to spawn, read from first line of file
-        getline(input, line);
-        istringstream numChildren(line);
-        numChildren >> children;
 
-        while(getline(input, line)) {
-            stringstream sstream;
-            sstream << line;
-            string s = sstream.str();
-            lines.push_back(s);
-        }
-    }
+    // TCP shit, talking to the routers
 
-    lines.pop_back();
-    for(auto v : lines) {
-        cout << v << "\n" << endl;
-    }
+    manager.Wait();
 
-    // As many pid's as children
-    pid_t pid[children];
-    // Spawn Children
-    int port;
-    for(int i=0; i<children; i++) {
-        port = (rand() % 999) + 9000;    // sets a random port between 9000 and 9999 for each
-        if ((pid[i] = fork()) < 0) {
-            perror("Fork failed");
-        } else if(pid[i] == 0) {
-            childFunction(fd, port);  // have child do something (write to file for now)
-            exit(0); // so children don't continue to fork
-        }
-    }
+
+
 }
 
 
@@ -70,13 +26,6 @@ int main(int argc, char * argv[]) {
     ss << argv[1];
     string in = ss.str();
     parentFunction(in);
-    // Wait for children to exit
-    int status;
-    // For each specific child
-    pid_t pidChild;
-    while ((pidChild = wait(&status)) > 0) {
-        cout << "Child with PID " << pidChild << " exited with status of " << status << endl;
-        children--;
-    }
+
     return 0;
 }
